fix: PR comments
