+++ frontend/src/features/admin-form/settings/SettingsEmailsPage.tsx
@@ -32,9 +32,8 @@ const FormEmailSectionContainer = ({
 }: FormEmailSectionContainerProps): JSX.Element => {
   if (settings.responseMode === FormResponseMode.Multirespondent) {
     return <MrfFormEmailSection isDisabled={isDisabled} settings={settings} />
-  } else {
-    return <FormEmailSection isDisabled={isDisabled} settings={settings} />
   }
+  return <FormEmailSection isDisabled={isDisabled} settings={settings} />
 }
 
 export const SettingsEmailsPage = (): JSX.Element => {
